refactor: Separate emoji functionality
